+++ ThunderstorePublish/CHANGELOG.md
@@ -1,2 +1,6 @@
 ## v1.0.0
 - Initial public version
+
+## v.1.0.1
+- Updated scanner item model
+- Fixed a bug when pocketed scanner lights weren't hidden